requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.65.0
openpyxl>=3.1.0
pyarrow>=14.0.0